refactor : 인터페이스 삭ㅈ
